Deep_Q_Network/requirements.txt
collections
numpy
gym
gymnasium
cv2
itertools
math
random
torch
torch.nn
torch.optim
torch.nn.functional
torchvision.transforms
time
csv
matplotlib.pyplot
pandas
re
os